Throttle requests by second/day
The Insightly API rate limits the number of requests per second to 5.
This class wraps the calls to the Insightly API
($this->throttle->attempt(Closure $callback)) keeping track of the
number of requests per day / second and sleeping when we are about to
hit the imposed rate limit. The daily limit exception is to be handled
at the bootstapping level of the application.