Get email ids from Insightly
Using our newly build Insightly API wrapper and Throttle classes, we
call the Insightly API to determine the total number of emails stored,
comparing it to the number of email IDs we already have on file and
fetching the remainder. This slightly reduces the number of API calls we
have to make.