Get individual emails
With the list of email IDs on file, we determine the emails we already
have on file, substract those from the email IDs file and fetch the
remainder. This approach is necessary in order to retrieve the full body
text of an email as that is not included in the /Emails.