Gracefully handle HTTP status 500 errors
The Insightly API contains a bug that causes a HTTP 500 error to be thrown for some emails we are trying to retrieve based on the email id.

The id is simply retrieved through the getEmails method and does infact exist. This commit simply writes the id to a file in case this happens.